refactor: change structure app for next js
