+++ __mocks__/wrapper.tsx
@@ -1,5 +1,5 @@
 import { Provider } from 'react-redux';
-import { store } from '../src/redux/store';
+import { store } from '../pages/api/store';
 import { ReactNode } from 'react';
 
 export function Wrapper(props: { children: ReactNode }) {
