+++ components/layout.module.css
@@ -20,6 +20,7 @@
 .searchField {
   display: flex;
   flex-wrap: wrap;
+  justify-content: center;
   gap: 10px;
   margin-bottom: 20px;
 }
