+++ components/layout.test.tsx
@@ -1,22 +1,8 @@
 import '@testing-library/jest-dom';
-import {
-  cleanup,
-  fireEvent,
-  render,
-  screen,
-  waitFor,
-} from '@testing-library/react';
-import {
-  Route,
-  RouterProvider,
-  createMemoryRouter,
-  createRoutesFromElements,
-} from 'react-router-dom';
-import App from '../../App';
-import ErrorPage from '../../error-page';
-import Card from '../card/Card';
-import { store } from '../../redux/store';
-import { Provider } from 'react-redux';
+import { cleanup, fireEvent, screen, waitFor } from '@testing-library/react';
+// import { store } from '../pages/api/store';
+// import { Provider } from 'react-redux';
+// import Layout from './layout';
 
 const mockUseNavigate = jest.fn().mockReturnValue('');
 jest.mock('react-router-dom', () => ({
@@ -30,24 +16,16 @@ jest.mock('react-router-dom', () => ({
     state: '',
   }),
 }));
-const routerSetting = createRoutesFromElements(
-  <Route path="/" element={<App />} errorElement={<ErrorPage />}>
-    <Route path="details/:name" element={<Card />} />
-  </Route>
-);
-const routerTest = createMemoryRouter(routerSetting, {
-  initialEntries: ['/'],
-});
 
 afterEach(cleanup);
 
 describe('Search component', () => {
   it('renders is correctly', () => {
-    render(
-      <Provider store={store}>
-        <RouterProvider router={routerTest} />
-      </Provider>
-    );
+    // render(
+    //   <Provider store={store}>
+    //     <Layout children={} />
+    //   </Provider>
+    // );
 
     const input = screen.getByRole('searchbox');
     const title = screen.getByText('Star Wars');
@@ -56,11 +34,11 @@ describe('Search component', () => {
     expect(title).toBeInTheDocument();
   });
   it('Verify that clicking the Search button saves the entered value to the local storage', () => {
-    render(
-      <Provider store={store}>
-        <RouterProvider router={routerTest} />
-      </Provider>
-    );
+    // render(
+    //   <Provider store={store}>
+    //     <Layout children={} />
+    //   </Provider>
+    // );
 
     const input = screen.getByRole('searchbox');
     fireEvent.change(input, { target: { value: 'Luke' } });
@@ -70,11 +48,11 @@ describe('Search component', () => {
   });
   it('retrieves the value from the local storage upon mounting', async () => {
     global.localStorage.setItem('searchWord', 'Luke');
-    render(
-      <Provider store={store}>
-        <RouterProvider router={routerTest} />
-      </Provider>
-    );
+    // render(
+    //   <Provider store={store}>
+    //     <Layout children={} />
+    //   </Provider>
+    // );
 
     const searchWord = screen.getByDisplayValue('LUKE');
