+++ components/list-results/ListResults.tsx
@@ -1,7 +1,6 @@
 import style from './listResults.module.css';
-// import { Outlet } from 'react-router-dom';
-import CardResults from './card-results/CardResults';
-import { useAppSelector } from '../../../redux/hooks';
+import { useAppSelector } from '../../redux/hooks';
+import Link from 'next/link';
 
 export default function ListResults() {
   const itemsPerPage = useAppSelector(
@@ -14,15 +13,16 @@ export default function ListResults() {
         <nav className={style.navigation}>
           {itemsPerPage?.length ? (
             <ul className={style.cards}>
-              {itemsPerPage.map((hero, index) => {
-                return <CardResults key={index} hero={hero} />;
-              })}
+              {itemsPerPage.map((hero, index) => (
+                <li key={index} className={style.link}>
+                  <Link href={`details/${hero.name}`}>{hero.name}</Link>
+                </li>
+              ))}
             </ul>
           ) : (
             <span>No Results</span>
           )}
         </nav>
-        <div className={style.detail}>{/* <Outlet /> */}</div>
       </div>
     </>
   );
