+++ pages/_app.tsx
@@ -1,14 +1,17 @@
 import { Provider } from 'react-redux';
 import '../styles/global.css';
 import type { AppProps } from 'next/app';
-import { wrapper } from '../src/redux/store';
+import { wrapper } from './api/store';
+import Layout from '../components/layout';
 
 export default function App({ Component, ...rest }: AppProps) {
   const { store, props } = wrapper.useWrappedStore(rest);
 
   return (
     <Provider store={store}>
-      <Component {...props.pageProps} />;
+      <Layout>
+        <Component {...props.pageProps} />;
+      </Layout>
     </Provider>
   );
 }
