+++ pages/detailes/card.module.css
@@ -1,5 +1,8 @@
 .cards {
   display: flex;
+  flex: 1;
+  padding: 2rem 4rem;
+  width: 100%;
   flex-wrap: wrap;
   gap: 10px;
   height: calc(100vh - 270px);
