+++ pages/index.test.tsx
@@ -1,13 +1,5 @@
 import '@testing-library/jest-dom';
-import { cleanup, render, screen } from '@testing-library/react';
-import {
-  Route,
-  RouterProvider,
-  createMemoryRouter,
-  createRoutesFromElements,
-} from 'react-router-dom';
-import App from '../../App';
-import Card from '../card/Card';
+import { cleanup, screen } from '@testing-library/react';
 
 const mockUseNavigate = jest.fn().mockReturnValue('');
 jest.mock('react-router-dom', () => ({
@@ -21,21 +13,11 @@ jest.mock('react-router-dom', () => ({
     state: '',
   }),
 }));
-const routerSetting = createRoutesFromElements(
-  <Route path="/" element={<App />}>
-    <Route path="details/:name" element={<Card />} />
-  </Route>
-);
-const routerTest = createMemoryRouter(routerSetting, {
-  initialEntries: ['/'],
-});
 
 afterEach(cleanup);
 
 describe('Results component', () => {
   it('renders is correctly', () => {
-    render(<RouterProvider router={routerTest} />);
-
     const input = screen.queryByAltText('loading');
 
     expect(input).not.toBeInTheDocument();
