+++ pages/index.tsx
@@ -1,11 +1,89 @@
-import Results from '../src/components/results/Results';
-import Search from '../src/components/search/Search';
+import style from './index.module.css';
+import ListResults from '../components/list-results/ListResults';
+import Pagination from '../components/pagination/Pagination';
+import { useEffect } from 'react';
+import { useAppSelector, useAppDispatch } from '../redux/hooks';
+import { updateItems } from '../redux/slices/itemsPerPageSlice';
+import {
+  getRunningQueriesThunk,
+  peopleApi,
+  useGetPeopleQuery,
+} from './api/peopleApi';
+import Spinner from '../components/spiner/Spinner';
+import { setLoadingMainPage } from '../redux/slices/loadingFlagsSlice';
+import { wrapper } from './api/store';
+import { GetServerSidePropsContext, PreviewData } from 'next';
+import { NextParsedUrlQuery } from 'next/dist/server/request-meta';
 
 export default function Page() {
+  const page: number = useAppSelector((state) => state.currentPage.pageNum);
+  const searchWord: string | null = useAppSelector(
+    (state) => state.searchWord.searchWord
+  );
+  const countPerPage = useAppSelector(
+    (state) => state.currentPage.countPerPage
+  );
+  const loadingMainPage = useAppSelector(
+    (state) => state.loadingFlags.mainPageLoading
+  );
+  const { data, isFetching } = useGetPeopleQuery({
+    page: page ?? 1,
+    searchWord: searchWord ?? '',
+  });
+
+  const dispatch = useAppDispatch();
+  useEffect(() => {
+    dispatch(updateItems(data?.results));
+  }, [data]);
+
+  useEffect(() => {
+    dispatch(setLoadingMainPage(isFetching));
+  }, [isFetching]);
+
+  const totalItems = data?.count;
+  if (!totalItems) return <Spinner />;
+  const totalPage = Math.ceil(totalItems / countPerPage);
+
   return (
-    <>
-      <Search />
-      <Results />
-    </>
+    <section className={style.results}>
+      {loadingMainPage ? (
+        <Spinner />
+      ) : (
+        <>
+          <ListResults />
+          <Pagination totalPage={totalPage} />
+        </>
+      )}
+    </section>
   );
 }
+
+export const getServerSideProps = wrapper.getServerSideProps(
+  (store) =>
+    async (
+      context: GetServerSidePropsContext<NextParsedUrlQuery, PreviewData>
+    ) => {
+      const page = context.query.page;
+      const searchWord = context.query.search;
+      if (!searchWord && typeof page === 'string') {
+        store.dispatch(
+          peopleApi.endpoints.getPeople.initiate({
+            page: Number(page),
+            searchWord: '',
+          })
+        );
+      }
+      if (typeof searchWord === 'string') {
+        store.dispatch(
+          peopleApi.endpoints.getPeople.initiate({
+            page: Number(page),
+            searchWord: searchWord,
+          })
+        );
+      }
+      await Promise.all(store.dispatch(getRunningQueriesThunk()));
+      // const res = await fetch('https://api.github.com/repos/vercel/next.js');
+      // const repo = await res.json();
+      return { props: {} };
+    }
+);
