+++ redux/hooks.ts
@@ -1,7 +1,7 @@
 import { useDispatch, useSelector } from 'react-redux';
 import type { TypedUseSelectorHook } from 'react-redux';
-import type { RootState, AppDispatch } from './store';
-import { peopleApi } from '../services/peopleApi';
+import type { RootState, AppDispatch } from '../pages/api/store';
+import { peopleApi } from '../pages/api/peopleApi';
 
 export const useAppSelector: TypedUseSelectorHook<RootState> = useSelector;
 export const useAppDispatch: () => AppDispatch = useDispatch;
