+++ redux/slices/searchWordSlice.ts
